Exposes buildOptionText property of Options component to Pagination component.
pageSizeBuildOptionText allows the user to define a method to overwrite the default buildOptionText method inside the Option component, allowing for custom text. The motivation for this is to put the data length as the last option in pageSizeOptions and then use this custom method to covert the display text to say "All".